--- a/bin/post_compile
+++ b/bin/post_compile
@@ -5,23 +5,16 @@
 set -euo pipefail
 
 # Make the current Git revision accessible at <site-root>/revision.txt
-echo "$SOURCE_VERSION" > ui/revision.txt
-
-# Create a `dist/` directory containing built/minified versions of the `ui/` assets.
-# Uses the node binaries/packages installed by the nodejs buildpack previously.
-yarn build
+echo "$SOURCE_VERSION" > build/revision.txt
 
 # Generate gzipped versions of files that would benefit from compression, that
 # WhiteNoise can then serve in preference to the originals. This is required
 # since WhiteNoise's Django storage backend only gzips assets handled by
-# collectstatic, and so does not affect files in the `dist/` directory.
-python -m whitenoise.compress dist
+# collectstatic, and so does not affect files in the `build/` directory.
+python -m whitenoise.compress build
 
 # Remove nodejs files to reduce slug size (and avoid environment variable
 # pollution from the nodejs profile script), since they are no longer
-# required once `yarn build` has run. The buildpack cache will still
-# contain them, so this doesn't slow down the next slug compile.
-rm -r .heroku/node/
-rm -r .heroku/yarn/
-rm -r .profile.d/nodejs.sh
-rm -r node_modules/
+# required after `yarn heroku-postbuild` has run. The buildpack cache will
+# still contain them, so this doesn't slow down the next slug compile.
+rm -r .heroku/node/ .heroku/yarn/ .profile.d/nodejs.sh node_modules/