--- a/setup.cfg
+++ b/setup.cfg
@@ -32,7 +32,7 @@ filterwarnings =
     error
     ignore::ImportWarning
     ignore::PendingDeprecationWarning
-    # WhiteNoise warns if either `treeherder/static/` or `dist/` do not exist at startup,
+    # WhiteNoise warns if either `treeherder/static/` or `build/` do not exist at startup,
     # however this is expected when running tests since Django collectstatic and yarn build
     # (which create those directories) typically aren't run apart from during deployments.
     ignore:No directory at.*:UserWarning:whitenoise.base