--- a/tests/ui/unit/init.js
+++ b/tests/ui/unit/init.js
@@ -1,17 +1,11 @@
 // Karma/webpack entry for tests
 
-import jQuery from 'jquery';
 // Manually import angular since angular-mocks doesn't do so itself
 import 'angular';
 import 'angular-mocks';
 import 'jasmine-jquery';
 import { configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-
-// Global variables are set here instead of with webpack.ProvidePlugin
-// because neutrino removes plugin definitions for karma runs:
-// https://github.com/mozilla-neutrino/neutrino-dev/issues/617
-window.jQuery = jQuery;
 
 configure({ adapter: new Adapter() });