--- a/ui/helpers/constants.js
+++ b/ui/helpers/constants.js
@@ -272,7 +272,7 @@ export const phTimeRangeValues = {
 export const phDefaultFramework = 'talos';
 
 export const phFrameworksWithRelatedBranches = [
-    1,  // talos
+    1, // talos
     10, // raptor
     11, // js-bench
     12, // devtools