--- a/ui/intermittent-failures/App.jsx
+++ b/ui/intermittent-failures/App.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { HashRouter, Route, Switch, Redirect } from 'react-router-dom';
+import { hot } from 'react-hot-loader';
 
 import MainView from './MainView';
 import BugDetailsView from './BugDetailsView';
@@ -57,4 +58,4 @@ class App extends React.Component {
   }
 }
 
-export default App;
+export default hot(module)(App);