--- a/ui/intermittent-failures/index.jsx
+++ b/ui/intermittent-failures/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { AppContainer } from 'react-hot-loader';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'font-awesome/css/font-awesome.css';
 import 'react-table/react-table.css';
@@ -9,16 +8,4 @@ import '../css/treeherder-global.css';
 import '../css/intermittent-failures.css';
 import App from './App';
 
-function load() {
-  render((
-    <AppContainer>
-      <App />
-    </AppContainer>
-  ), document.getElementById('root'));
-}
-
-load();
-
-if (module.hot) {
-  module.hot.accept('./App', () => load(App));
-}
+render(<App />, document.getElementById('root'));