--- a/ui/job-view/index.jsx
+++ b/ui/job-view/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { AppContainer } from 'react-hot-loader';
 
 // Vendor Styles
 import 'bootstrap/dist/css/bootstrap.min.css';
@@ -25,14 +24,4 @@ import '../css/treeherder-loading-overlay.css';
 
 import App from './App';
 
-const load = () => render((
-  <AppContainer>
-    <App />
-  </AppContainer>
-), document.getElementById('root'));
-
-if (module.hot) {
-  module.hot.accept('./App', load);
-}
-
-load();
+render(<App />, document.getElementById('root'));