--- a/ui/login-callback/LoginCallback.jsx
+++ b/ui/login-callback/LoginCallback.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import Icon from 'react-fontawesome';
+import { hot } from 'react-hot-loader';
 
 import AuthService from '../shared/auth/AuthService';
 import { webAuth, parseHash } from '../helpers/auth';
@@ -74,4 +75,4 @@ class LoginCallback extends React.PureComponent {
   }
 }
 
-export default LoginCallback;
+export default hot(module)(LoginCallback);