--- a/ui/test-view/index.jsx
+++ b/ui/test-view/index.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
-import { AppContainer } from 'react-hot-loader';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'font-awesome/css/font-awesome.css';
 
@@ -9,16 +8,8 @@ import '../css/treeherder-test-view.css';
 import { store, actions } from './redux/store';
 import App from './ui/App';
 
-const load = () => render((
-  <AppContainer>
-    <Provider store={store}>
-      <App actions={actions} />
-    </Provider>
-  </AppContainer>
+render((
+  <Provider store={store}>
+    <App actions={actions} />
+  </Provider>
 ), document.getElementById('root'));
-
-if (module.hot) {
-  module.hot.accept('./ui/App', load);
-}
-
-load();