--- a/ui/test-view/ui/App.jsx
+++ b/ui/test-view/ui/App.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { hot } from 'react-hot-loader';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 
 import Navigation from './Navigation';
@@ -41,4 +42,4 @@ App.defaultProps = {
   location: null,
 };
 
-export default App;
+export default hot(module)(App);