--- a/ui/userguide/App.jsx
+++ b/ui/userguide/App.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { hot } from 'react-hot-loader';
 
 import UserGuideHeader from './UserGuideHeader';
 import UserGuideBody from './UserGuideBody';
@@ -14,4 +15,4 @@ const App = () => (
   </div>
 );
 
-export default App;
+export default hot(module)(App);