--- a/ui/userguide/index.jsx
+++ b/ui/userguide/index.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { AppContainer } from 'react-hot-loader';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 import 'font-awesome/css/font-awesome.css';
@@ -11,14 +10,4 @@ import '../css/treeherder-job-buttons.css';
 
 import App from './App';
 
-const load = () => render((
-  <AppContainer>
-    <App />
-  </AppContainer>
-), document.getElementById('root'));
-
-if (module.hot) {
-  module.hot.accept('./App', load);
-}
-
-load();
+render(<App />, document.getElementById('root'));