Neutrino controls our frontend linting, transpilation, source-maps, testing, dev-server and optimisation of production builds.

Highlights of the upgrade are:

* Major version updates to the individual tools within (such as webpack, Babel and ESLint), significantly improving performance, fixing transpilation/minification correctness bugs, adding support for newer ECMAScript features, and increasing linter coverage.
* Hot reloading in the dev server now works for all entry-points and not just the jobs view, shortening the feedback cycle.
* Reduced bundle size due to webpack 4's [tree shaking](https://webpack.js.org/guides/tree-shaking/), scope hoisting, automatic shared/vendor code chunk splitting (no need for the manually maintained 'vendor' list).
* CSS is now extracted out of JS, which improves performance, reduces bundle size and prevents the initial white flash of un-styled content.
* Support for [dynamic imports/code splitting](https://webpack.js.org/guides/code-splitting/) (needed for [bug 1502192](https://bugzilla.mozilla.org/show_bug.cgi?id=1502192)).
* Support for Jest via a new Jest preset (unblocks [bug 1364045](https://bugzilla.mozilla.org/show_bug.cgi?id=1364045)).
* Support for [public class field declarations](https://github.com/tc39/proposal-class-fields#field-declarations) (unblocks [bug 1480166](https://bugzilla.mozilla.org/show_bug.cgi?id=1480166)).
* Improved source-maps (increases the quality of production exception trace-backs and fixes several debugger breakpoint bugs).
* Reduced amount of custom configuration required for our fairly complex frontend needs, reducing maintenance burden and allowing for easier future Neutrino upgrades.

In addition this PR:

* Fixes the WhiteNoise `immutable_file_test()` regex, so that it now correctly enables browser caching of images, fonts and source maps.
* Enables webpack-dev-server's overlay feature, which displays any compilation errors in the browser, saving having to switch back to the console (this can be enabled for warnings too if desired).
* Enables webpack-dev-server's automatic browser-opening feature, which saves having to manually navigate to `localhost:5000` after running `yarn start`.
* Switches Karma tests to run Firefox in headless mode, reducing the workflow disruption when running `yarn test`.
* Uses the new webpack `performance` option to enable maximum asset file size thresholds, to help prevent bundle-size regressions.
* Rewrites the `package.json` script commands so that they now work correctly on Windows, even when setting environment variables.

Performance comparison:

| Workflow | Before | After |
|:--------:|:------:|:-----:|
| Local `yarn build` | 2m34s | Cached: 23s <br/> Uncached: 58s |
| Local `yarn start` | 34.5s | Cached: 13.6s <br/> Uncached: 31.3s |
| Local `yarn test` | 61.5s | Cached: 19.8s <br/> Uncached: 22.0s |
| Local `yarn lint` | Cached: 3.8s <br/> Uncached: 13.7s | Cached: 1.8s <br/> Uncached: 13.4s |
| TravisCI end-to-end time | ~9 minutes | ~6 minutes |
| Heroku deploy end-to-end time | ~14 minutes | ~9 minutes |