@camd 2018-11-01
Hmm..  @edmorley : I take that back, though it may be user error or something.  :)  When I make a mod in one of the react files, the hotloader reloads, but gives a lot of errors on the console about:

    Warning: Failed prop type: The prop `notify` is marked as required in `PinnedJobsClass`, but its value is `undefined`.