@camd 2018-11-01
I get a ton of those error messages.  It appears they're all values that are coming from the ``Context`` objects.  It seems benign and only happens on the FIRST hot reload change.  After that, things hot reload without noise.