@edmorley 2018-11-02
> I get a ton of those error messages. It appears they're all values that are coming from the `Context` objects. It seems benign and only happens on the FIRST hot reload change. After that, things hot reload without noise.

Yeah I get them too - it's likely a bug with react-hot-loader and the context API.
I've filed gaearon/react-hot-loader#1094 with a reduced testcase.

Since they are mostly benign (and not a regression with Neutrino/webpack/... themselves) are you fine if we merge this anyway?