@camd 2018-11-02
Yeah, that's totally fine by me.  The benefits of this upgrade are huge and totally worth it.  :)